pyproject.toml
[project]
name = "quriya-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt>=5.0.0",
    "celery[redis]>=5.6.0",
    "fastapi[standard]>=0.122.0",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "resend>=2.19.0",
    "sqlalchemy>=2.0.44",
]
